requirements.txt
#deprecated use reruirements-dev.txt instead
Pillow
PyQt5
